+++ CHANGELOG.md
@@ -1,23 +1,40 @@
-# rocDecode changelog
+# Changelog for rocDecode
 
-Documentation for rocDecode is available at
-[https://rocm.docs.amd.com/projects/rocDecode/en/latest/](https://rocm.docs.amd.com/projects/rocDecode/en/latest/)
+Full documentation for rocDecode is available at [https://rocm.docs.amd.com/projects/rocDecode/en/latest/](https://rocm.docs.amd.com/projects/rocDecode/en/latest/)
+
+## rocDecode 0.8.0 for ROCm 6.3
+
+### Changed
+
+* Clang is now the default CXX compiler.
+* The new minimum supported version of va-api is 1.16.
+* New build and runtime options have been added to the `rocDecode-setup.py` setup script.
+
+### Removed
+
+* Make tests have been removed. CTEST is now used for both Make tests and package tests.
+* `mesa-amdgpu-dri-drivers` has been removed as a dependency on RHEL and SLES.
+
+### Resolved issues
+
+* Fixed a bug in the size of output streams in the `videoDecodeBatch` sample.
 
 ## rocDecode 0.7.0
 
-### Additions
+### Added
 
 * Clang - Default CXX compiler
+* Parser - Add new API rocDecParserMarkFrameForReuse()
 
-### Optimizations
+### Optimized
 
 * Setup Script - Build and runtime install options
 
-### Changes
+### Changed
 
 * CTest - Core tests for make test and package test
 
-### Fixes
+### Resolved issues
 
 * Sample - Bugfix for videoDecodeBatch
 
@@ -32,7 +49,6 @@ Documentation for rocDecode is available at
   * amdgpu-core - `1:6.2.60200-2009582`
 * libva-dev - `2.7.0-2` / `2.14.0-1`
 * mesa-amdgpu-va-drivers - `1:24.2.0.60200-2009582`
-* mesa-amdgpu-dri-drivers - `24.1.0.60200`
 * FFmpeg - `4.2.7` / `4.4.2-0`
 * rocDecode Setup Script - `V2.2.0`